fix(composables): fix composables structure
improve composables structure by exporting all of the divided  composables from index.ts.

#158
