+++ src/components/MenuDropdown/LuiMenuDropdown.vue
@@ -9,8 +9,7 @@ export default {
 import { computed, nextTick, reactive, ref, useSlots } from 'vue'
 import type { PropType } from 'vue'
 import LuiButton from '../Button/LuiButton.vue'
-import { useOutsideClick } from '../../composables/useOutsideClick'
-import { useProperPosition } from '../../composables/useProperPosition'
+import { useOutsideClick, useProperPosition } from '../../composables'
 import { useId } from '../../utils/useId'
 import type { Block, Color, Filter, Rounded, Size, Variant } from '@/globals/types'
 import type { TwClassInterface } from '@/globals/interfaces'
